requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.2.0
python-multipart==0.0.6
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.24.3
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
